Laugh & Baka Command

+++ src/events/bot/guildDelete.ts
@@ -16,13 +16,13 @@ export default class GuildDeleteEvent extends Event {
     }
 
     async run(guild: Guild): Promise<void> {
-        await ReactionRoles.findOneAndDelete({ guildID: guild.id }).catch(() => void 0);
+        if(!guild.name) return;
+        await ReactionRoles.deleteMany({ guildID: guild.id }).catch(() => void 0);
         await Welcomes.findOneAndDelete({ guildID: guild.id }).catch(() => void 0);
         await Servers.findOneAndDelete({ guildID: guild.id }).catch(() => void 0);
         await Leaves.findOneAndDelete({ guildID: guild.id }).catch(() => void 0);
-        await Embeds.findOneAndDelete({ guildID: guild.id }).catch(() => void 0);
-        await Tags.findOneAndDelete({ guildID: guild.id }).catch(() => void 0);
-
+        await Embeds.deleteMany({ guildID: guild.id }).catch(() => void 0);
+        await Tags.deleteMany({ guildID: guild.id }).catch(() => void 0);
         const guilds = (await this.client.shard!.fetchClientValues('guilds.cache.size')).reduce((acc, guildCount) => acc + guildCount, 0);
         const channel = await this.client.channels.fetch(process.env.SERVERS_CHANNEL) as TextChannel;
         const owner = await this.client.users.fetch(guild.ownerID).catch(() => null);
